Move hw_core::state::tests to separate file.
